Renamed to EjbDiagnosticsParticipant, Refactored EJB diagnostics to use ICompilationUnit#getAllTypes() pattern and getMatchedJavaElementNames() utility
