+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/plugin.xml
@@ -293,7 +293,7 @@
    
    <!-- EJB -->
    <extension point="org.eclipse.lsp4jakarta.jdt.core.javaFeatureParticipants">
-      <diagnostics class="org.eclipse.lsp4jakarta.jdt.internal.ejb.SessionBeanDiagnosticsParticipant" />
+      <diagnostics class="org.eclipse.lsp4jakarta.jdt.internal.ejb.EjbDiagnosticsParticipant" />
       <codeAction kind="quickfix"
                   targetDiagnostic="jakarta-ejb#MissingPublicNoArgConstructor"
                   class="org.eclipse.lsp4jakarta.jdt.internal.ejb.InsertPublicNoArgConstructorQuickFix" />
